# OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: Project Management Orchestrator - Revised
  version: 2.0.0
  description: |
    Unified API with proper organization isolation and role-based access control.
    
    **Authentication**: All endpoints require Bearer token except /register and /login
    **Authorization**: Enforced at endpoint level based on user role
    **Data Isolation**: Users can only access data from their organization
servers:
  - url: http://localhost:9000
    description: Development server
  - url: https://api.example.com
    description: Production server

# ============================================
# AUTHENTICATION & ORGANIZATION MANAGEMENT
# ============================================
paths:
...
  /api/v1/organizations:
    post:
      tags: [Organizations]
      summary: Create organization
      description: |
        Creates new organization and makes current user the admin.
        ⚠️ User must not already be in an organization.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Organization created, user is now admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  owner_id:
                    type: string
                    format: uuid
        '400':
          description: User already in organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Organizations]
      summary: Get user's organization
      description: Returns the organization the user belongs to
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                type: object
        '404':
          description: User not in any organization
...
  /api/v1/projects:
    post:
      tags: [Projects]
      summary: Create project
      description: |
        ⚠️ Only admin or manager can create projects.
        Creates project in user's organization.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                lab_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Optional lab link
      responses:
        '201':
          description: Project created
        '403':
          description: Only admin or manager can create projects
        '400':
          description: Lab not in organization (if lab_id provided)

    get:
      tags: [Projects]
      summary: List projects in user's organization
      security:
        - bearerAuth: []
      parameters:
        - name: lab_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by lab
      responses:
        '200':
          description: List of projects
  /api/v1/projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project details
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details with tasks
        '403':
          description: Not authorized to view this project
  /api/v1/projects/{project_id}/tasks:
    post:
      tags: [Projects]
      summary: Create task in project
      description: |
        ⚠️ Only admin or manager can create tasks.
        Assigns task to org member.
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                assignee_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Must be org member
                due_date:
                  type: string
                  format: date
                status:
                  type: string
                  enum: [To Do, In Progress, Done]
                  default: To Do
      responses:
        '201':
          description: Task created
        '403':
          description: Only admin or manager can create tasks
        '400':
          description: Assignee not in organization

    get:
      tags: [Projects]
      summary: List tasks in project
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of tasks
  /api/v1/tasks/{task_id}:
    patch:
      tags: [Projects]
      summary: Update task
      description: Assignee or creator can update task
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [To Do, In Progress, Done]
                progress_percentage:
                  type: integer
                  minimum: 0
                  maximum: 100
      responses:
        '200':
          description: Task updated
        '403':
          description: Not authorized to update this task

# ============================================
# ACTIVITY LOGGING (WorkPulse Service)
# ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string